fixes creacion de componente dragable
